finalize: implement reconfigure-ports substep
The reconfigure-ports substep has been fully reimplemented, and the CLI
command to use it is now `gpupgrade finalize`.

The core of the reconfigure-ports implementation is
ClonePortsFromCluster, which sets up the master gp_segment_configuration
on the new cluster to match the old one. Ideally it would also perform

    ALTER SYSTEM SET port

on the new master, but bugs in gpstart and gpstop currently prevent us
from doing that, so for now we keep the (brittle) sed replacement
strategy for postgresql.conf that was in place previously.

The new end-to-end test sets up a new "post-upgrade" cluster and runs
finalize on it. The new testutils/dump_config utility helps with this by
writing the manually-constructed cluster's JSON configuration to disk.

The test can't actually run in the CI yet due to two bugs:

- kill_agents doesn't kill agents on all hosts in the cluster, and
- we expect gpupgrade_* to be in $GPHOME/bin instead of looking for it
  on the PATH. (This is the same bug that prevents us from running the
  execute end-to-end test.)

delete_cluster has been moved to test/helpers, since it's used from two
places now. It also performs the port lookup via PID file, instead of
the environment, since we end up changing ports during execution.

The fact that we ignore error codes coming from gpinitsystem and
gpdeletesystem continues to be a travesty.

Co-authored-by: David Krieger <dkrieger@pivotal.io>
Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>
Co-authored-by: Kalen Krempely <kkrempely@pivotal.io>
Co-authored-by: Shoaib Lari <slari@pivotal.io>