gpupgrade: move executables from GPHOME to PATH
We don't want a separate version of gpupgrade per GPDB installation;
there should be only one best version to use. So look for it on the
PATH. By default the Makefile will install under the /usr/local prefix,
but that can be overridden by explicitly specifying a different PREFIX
during `make install`. (Some of the internal variables have been renamed
to make that "standard" variable name available.)

This will also help fix `make check` tests that want to test locally
built binaries and not ones that are installed, because they can simply
override the PATH to point wherever they want.